Move helpers out
