+++ crypto_verify.js
@@ -1,11 +1,7 @@
 /* eslint-disable camelcase */
-const assert = require('nanoassert')
-
 module.exports = {
   crypto_verify_16,
-  crypto_verify_32,
-  sodium_memcmp,
-  sodium_is_zero
+  crypto_verify_32
 }
 
 function vn (x, xi, y, yi, n) {
@@ -14,22 +10,15 @@ function vn (x, xi, y, yi, n) {
   return (1 & ((d - 1) >>> 8)) - 1
 }
 
+// Make non enumerable as this is an internal function
+Object.defineProperty(module.exports, 'vn', {
+  value: vn
+})
+
 function crypto_verify_16 (x, xi, y, yi) {
   return vn(x, xi, y, yi, 16)
 }
 
 function crypto_verify_32 (x, xi, y, yi) {
   return vn(x, xi, y, yi, 32)
-}
-
-function sodium_memcmp (a, b) {
-  assert(a.byteLength === b.byteLength, 'buffers must be the same size')
-
-  return vn(a, 0, b, 0, a.byteLength) === 0
-}
-
-function sodium_is_zero (arr) {
-  var d = 0
-  for (let i = 0; i < arr.length; i++) d |= arr[i]
-  return d === 0
 }
